refactor: move HTTP_METHODS enum to api.ts for better organization
- Removed HTTP_METHODS enum from constants.ts and added it to api.ts for improved structure and clarity in API-related utilities.
